### 2021-05-14 #3288 @whedon
[PRE REVIEW]: iharm3D: Vectorized General Relativistic Magnetohydrodynamics
**Submitting author:** @bprather (<a href="http://orcid.org/0000-0002-0393-7734">Ben S. Prather</a>)
**Repository:** <a href="https://github.com/AFD-Illinois/iharm3d/" target ="_blank">https://github.com/AFD-Illinois/iharm3d/</a>
**Version:** v3.6-alpha
