### 2021-05-14 #3288 @danielskatz
@bprather - if you have time, you could work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. You could also remove any entries from your bib file that are not actually referenced, if there are any, which will make this process easier.  Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
