### 2021-05-14 60c7a82
Clean up reference management and make sure more references have DOIs
