+++ paper/paper.bib
@@ -1,309 +1,60 @@
 
-@article{porth_black_2017,
-	title = {The {Black} {Hole} {Accretion} {Code}},
-	volume = {4},
-	issn = {2197-7909},
-	url = {http://arxiv.org/abs/1611.09720},
-	doi = {10.1186/s40668-017-0020-2},
-	abstract = {We present the black hole accretion code (BHAC), a new multidimensional general-relativistic magnetohydrodynamics module for the MPI-AMRVAC framework. BHAC has been designed to solve the equations of ideal general-relativistic magnetohydrodynamics in arbitrary spacetimes and exploits adaptive mesh refinement techniques with an efficient block-based approach. Several spacetimes have already been implemented and tested. We demonstrate the validity of BHAC by means of various one-, two-, and three-dimensional test problems, as well as through a close comparison with the HARM3D code in the case of a torus accreting onto a black hole. The convergence of a turbulent accretion scenario is investigated with several diagnostics and we find accretion rates and horizon-penetrating fluxes to be convergent to within a few percent when the problem is run in three dimensions. Our analysis also involves the study of the corresponding thermal synchrotron emission, which is performed by means of a new general-relativistic radiative transfer code, BHOSS. The resulting synthetic intensity maps of accretion onto black holes are found to be convergent with increasing resolution and are anticipated to play a crucial role in the interpretation of horizon-scale images resulting from upcoming radio observations of the source at the Galactic Center.},
+@article{porth_event_2019,
+	title = {The {Event} {Horizon} {General} {Relativistic} {Magnetohydrodynamic} {Code} {Comparison} {Project}},
+	url = {http://arxiv.org/abs/1904.04923},
+	doi = {10.3847/1538-4365/ab29fd},
+	abstract = {Recent developments in compact object astrophysics, especially the discovery of merging neutron stars by LIGO, the imaging of the black hole in M87 by the Event Horizon Telescope (EHT) and high precision astrometry of the Galactic Center at close to the event horizon scale by the GRAVITY experiment motivate the development of numerical source models that solve the equations of general relativistic magnetohydrodynamics (GRMHD). Here we compare GRMHD solutions for the evolution of a magnetized accretion flow where turbulence is promoted by the magnetorotational instability from a set of nine GRMHD codes: Athena++, BHAC, Cosmos++, ECHO, H-AMR, iharm3D, HARM-Noble, IllinoisGRMHD and KORAL. Agreement between the codes improves as resolution increases, as measured by a consistently applied, specially developed set of code performance metrics. We conclude that the community of GRMHD codes is mature, capable, and consistent on these test problems.},
+	urldate = {2019-09-11},
+	journal = {arXiv:1904.04923 [astro-ph, physics:gr-qc]},
[…]
