### 2021-05-14 #3288 @bprather
@eloisabentivegna Thanks for taking this on!

@danielskatz is there something I'm missing activating whedon?  In any case I've taken another look at the references and I believe all the missing DOIs should be fixed in 'joss-paper-branch'
