### 2021-05-14 #3288 @danielskatz
I'm not sure why this didn't work - let's just try it again.  If it fails, there's perhaps a problem in the bib file that is causing it to fail silently.  If this happens, we might try to generate the PDF and see if that tells us the error
