### 2021-05-18 #3288 @eloisabentivegna
@bgiacoma, I will send you an email with all the details. In the meantime, you can take a look here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
