### 2021-05-19 #3288 @bgiacoma
@eloisabentivegna I received the email and I also had a look at the guidelines. Are the authors the same ones listed as contributors in the github repository of the code? Just to be sure that there are no conflict of interests.
