### 2021-05-19 #3288 @eloisabentivegna
@bgiacoma, you can take a look at the [paper draft](https://github.com/openjournals/joss-papers/blob/joss.03288/joss.03288/10.21105.joss.03288.pdf). The author list usually coincides with the set of code authors, but sometimes there are reasons why it does not. The author list in the paper is the one you should keep in mind (although check the [JOSS COI policy](https://github.com/openjournals/joss/blob/master/COI.md) for sources of conflict that go beyond direct collaboration with the individual authors).
