### 2021-05-24 #3288 @felker
Unfortunately I cannot, but I would suggest to the authors to change the reference for Athena++: https://ui.adsabs.harvard.edu/abs/2020ApJS..249....4S/abstract

See details here: https://github.com/PrincetonUniversity/athena
