### 2021-05-24 90bebf5
Add citation to Stone et al 2020 for Athena++ GRMHD code
