### 2021-06-05 #3336 @whedon
[REVIEW]: iharm3D: Vectorized General Relativistic Magnetohydrodynamics
**Submitting author:** @bprather (<a href="http://orcid.org/0000-0002-0393-7734">Ben S. Prather</a>)
**Repository:** <a href="https://github.com/AFD-Illinois/iharm3d/" target ="_blank">https://github.com/AFD-Illinois/iharm3d/</a>
**Version:** v3.6-alpha
