### 2021-06-21 0c9d9a0
Revert "Add basic version of a floor call after reconstruction"
This reverts commit 8b97ec2a764bed1e716aea0d949f7a90f2c473ac.
