+++ core/fluxes.c
@@ -116,32 +116,25 @@ void lr_to_flux(struct GridGeom *G, struct FluidState *Sr,
 
   // Properly offset left face
   // These are un-macro'd to bundle OpenMP thread tasks rather than memory accesses
-  double floor = 1.e-7;
   PLOOP {
     if (dir == 1) {
 #pragma omp parallel for collapse(2)
-      ZSLOOP_REVERSE(-1, N3, -1, N2, -1, N1) {
-        Sl->P[ip][k][j][i] = (ip < 2) ? MY_MAX(Sl->P[ip][k][j][i - 1], floor) : Sl->P[ip][k][j][i - 1];
-        Sr->P[ip][k][j][i] = (ip < 2) ? MY_MAX(Sr->P[ip][k][j][i], floor) : Sr->P[ip][k][j][i];
-      }
+      ZSLOOP_REVERSE(-1, N3, -1, N2, -1, N1)
+        Sl->P[ip][k][j][i] = Sl->P[ip][k][j][i - 1];
     } else if (dir == 2) {
 #pragma omp parallel for collapse(2)
       for (int k = (N3) + NG; k >= (-1) + NG; k--) {
         for (int i = (N1) + NG; i >= (-1) + NG; i--) {
-          for (int j = (N2) + NG; j >= (-1) + NG; j--) {
-            Sl->P[ip][k][j][i] = (ip < 2) ? MY_MAX(Sl->P[ip][k][j - 1][i], floor) : Sl->P[ip][k][j - 1][i];
-            Sr->P[ip][k][j][i] = (ip < 2) ? MY_MAX(Sr->P[ip][k][j][i], floor) : Sr->P[ip][k][j][i];
-          }
+          for (int j = (N2) + NG; j >= (-1) + NG; j--)
+            Sl->P[ip][k][j][i] = Sl->P[ip][k][j - 1][i];
         }
       }
     } else if (dir == 3) {
 #pragma omp parallel for collapse(2)
       for (int j = (N2) + NG; j >= (-1) + NG; j--) {
         for (int i = (N1) + NG; i >= (-1) + NG; i--) {
-          for (int k = (N3) + NG; k >= (-1) + NG; k--) {
-            Sl->P[ip][k][j][i] = (ip < 2) ? MY_MAX(Sl->P[ip][k - 1][j][i], floor) : Sl->P[ip][k - 1][j][i];
-            Sr->P[ip][k][j][i] = (ip < 2) ? MY_MAX(Sr->P[ip][k][j][i], floor) : Sr->P[ip][k][j][i];
-          }
+          for (int k = (N3) + NG; k >= (-1) + NG; k--)
+            Sl->P[ip][k][j][i] = Sl->P[ip][k - 1][j][i];
         }
       }
     }
