### 2021-06-29 #3336 @bgiacoma
@eloisabentivegna I cannot find the pdf of the paper. Sorry, I'm new to this system.
