### 2021-06-29 AFD-Illinois/iharm3d#30 @bgiacoma
missing example usage
from the README it is not clear to me how to run the torus problem that is automatically generated when compiling. Can you add more information? Can it be run on a single core on a laptop?
