### 2021-06-29 #3336 @bgiacoma
@bprather I added an issue about the README file (issue #30) on the github repository of the code
[https://github.com/AFD-Illinois/iharm3d/issues/30#issue-932686829](https://github.com/AFD-Illinois/iharm3d/issues/30#issue-932686829)
