### 2021-06-29 #3336 @eloisabentivegna
> FYI, @eloisabentivegna - when you start a review, its a good idea to check to make sure there's a paper produced, and if the source is in a branch, do this manually...

Thanks, Dan! I usually do this in pre-review (although you beat me to it there too this time... :) ). Is it better to do it (or do it again) in the review issue?
