### 2021-07-01 #3336 @cpalenzuela
@eloisabentivegna I cannot edit my review checklist . I made sure I am logged in, but in the step 2 (Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations) I got the following error message "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account."
