### 2021-07-05 be79509
steps to run the FM torus
