+++ README.md
@@ -52,6 +52,26 @@ remove just the object and executable files, preserving any customizations in `b
 Full details of production runs on larger machines e.g. Stampede2 are in `script/submit/checklist.txt` in this repository, along
 with job submission scripts for SLURM in the TACC environment, adaptable for a lot of SLURM machines.
 
+## Running a Fishbone-Moncrief torus
+
+The [Fishbone-Moncrief](https://doi.org/10.1086/154565)(FM) torus is the ubiquitous initial condition for modelling compact radio sources such as M87 and SgrA*. The FM problem can be simulated on `iharm3d` by using the `-PROB=torus` flag while making the program, and specifying the problem-specific parameters (compile-time and run-time) in `parameters.h` (in the build archive) and an additional parameter file which, by default `iharm3d` assumes to be named `param.dat`. The following steps outline the commands to compile and execute the problem. They presuppose that all the necessary dependencies (eg: OpenMP, MPI, phdf5, GSL) are installed and the directory variables and flags in `makefile` are pointing to them correctly.
+
+1. Invoke the make command from the output directory, 
+
+```bash
+$ make -f IHARM3D_DIRECTORY/makefile PROB=torus
+```
+   where IHARM3D_DIRECTORY is the path to your local `iharm3d` repository that contains the makefile. The output directory is where, as explained in the section above, the harm executable is created along with the `build_archive`. `build_archive` contains the source files necessary to run `iharm3d` along with the problem-specific compile-time parameter file, `parameters.h` and problem initialization file, `problem.c`.
+
+2. Modify compile-time parameters in `build_archive/parameters.h`. These typically include (i) the grid size `NiTOT`; (ii) number of MPI ranks `NiCPU`; (iii) density and internal energy floors, `BSQORHOMAX`, `UORHOMAX`, `BSQOUMAX`; (iv) the reconstruction scheme `RECONSTRUCTION`. NOTE: If you're running `iharm3d` on your local system, it is recommended that the FM problem is run at a low resolution or a 2D problem is executed (set `N3TOT` to 1).
+
+3. If the compile-time parameters have been modified or the C code in any of the source files in `build_archive` has been edited, the harm executable must be remade with the same command as in (1) from the output directory.
+
+4. Copy any of the parameter files located at `IHARM3D_DIRECTORY/prob/torus/` labelled param_sane.dat or param_mad.dat to the output directory and rename the file as `param.dat`. This contains the runtime parameters for the FM torus (eg: duration of run, domain size, output file cadence, fluid properties, FM torus size, FMKS grid geometry). NOTE: It is again recommended to set `tf` to a reasonable value if you're running the problem on your local computer.
+
+5. Submitting the run: Once the parameter
+
+
 ## Hacking
 Notes that may save you time in reading the source code:
 * Grid coordinates match physical coordinates i => x^1, j => x^2, k => x^3.  However, they are indexed backward
