### 2021-07-06 efa4445
incorporated suggestions from people, fixed typos
