### 2021-07-06 AFD-Illinois/iharm3d#30 @bprather
Hi, thanks for catching this!  The knowledge was floating around e.g. script/submit/checklist.txt and proprietary places, and just never got collated in the README.

I think MR #31 should be enough to get you started -- thanks @vedantdhruv96!

If you're interested in plotting: the scripts in script/analysis are written for 3D simulations mostly, and may need to be modified to plot 2D output. 3D simulations can be plotted with e.g. `python script/analysis/movie.py simplest dumps/`
For 2D simulations, it might be easier to use the newer scripts from [pyHARM](https://github.com/AFD-Illinois/pyHARM) -- with e.g. `python scripts/movie.py log_rho_poloidal /path/to/dumps/`.  The example notebook there has some other ways to play with the data directly.
