### 2021-07-07 AFD-Illinois/iharm3d#30 @bgiacoma
I was able to run the Fishbone-Moncrief torus following the new README instructions. I had to play a bit with the grid setup in build_archive/parameters.h in order to get something small to run, but not too small that the code complained (at the end 128*32*32).
Instead I have problems running the movie.py script since it fails while looking for 'eht_out.p' (since it does not exist in the dumps directory)
