### 2021-07-09 9cbfc0f
Scripts: Fix movies and 2D run plotting by backporting some fixes from pyHARM
