+++ script/analysis/movie.py
@@ -70,7 +70,7 @@ def plot(n):
 
   if movie_type == "simplest":
     # Simplest movie: just RHO
-    ax_slc = [plt.subplot(1,2,1), plt.subplot(1,2,2)]
+    ax_slc = [fig.add_subplot(1,2,1), fig.add_subplot(1,2,2)]
     bplt.plot_xz(ax_slc[0], geom, np.log10(dump['RHO']),
                      label="", vmin=rho_l, vmax=rho_h, window=window,
                      xlabel=False, ylabel=False, xticks=False, yticks=False,
@@ -86,21 +86,21 @@ def plot(n):
   elif movie_type == "simpler":
     # Simpler movie: RHO and phi
     gs = gridspec.GridSpec(2, 2, height_ratios=[6, 1], width_ratios=[16,17])
-    ax_slc = [fig.subplot(gs[0,0]), fig.subplot(gs[0,1])]
-    ax_flux = [fig.subplot(gs[1,:])]
+    ax_slc = [fig.add_subplot(gs[0,0]), fig.add_subplot(gs[0,1])]
+    ax_flux = [fig.add_subplot(gs[1,:])]
     bplt.plot_slices(ax_slc[0], ax_slc[1], geom, dump, np.log10(dump['RHO']),
                      label=r"$\log_{10}(\rho)$", vmin=rho_l, vmax=rho_h, window=window,
                      overlay_field=False, cmap='jet')
-    bplt.diag_plot(ax_flux[0], diag, 'phi_b', dump['t'], ylabel=r"$\phi_{BH}$", logy=LOG_PHI, xlabel=False)
+    bplt.diag_plot(ax_flux[0], diag, 'Phi_b', dump['t'], ylabel=r"$\phi_{BH}$", logy=LOG_PHI, xlabel=False)
   elif movie_type == "simple":
     # Simple movie: RHO mdot phi
     gs = gridspec.GridSpec(3, 2, height_ratios=[4, 1, 1])
-    ax_slc = [fig.subplot(gs[0,0]), fig.subplot(gs[0,1])]
-    ax_flux = [fig.subplot(gs[1,:]), fig.subplot(gs[2,:])]
+    ax_slc = [fig.add_subplot(gs[0,0]), fig.add_subplot(gs[0,1])]
+    ax_flux = [fig.add_subplot(gs[1,:]), fig.add_subplot(gs[2,:])]
     bplt.plot_slices(ax_slc[0], ax_slc[1], geom, dump, np.log10(dump['RHO']),
                      label=r"$\log_{10}(\rho)$", vmin=rho_l, vmax=rho_h, window=window, cmap='jet')
-    bplt.diag_plot(ax_flux[0], diag, 'mdot', dump['t'], ylabel=r"$\dot{M}$", logy=LOG_MDOT)
-    bplt.diag_plot(ax_flux[1], diag, 'phi_b', dump['t'], ylabel=r"$\phi_{BH}$", logy=LOG_PHI)
+    bplt.diag_plot(ax_flux[0], diag, 'Mdot', dump['t'], ylabel=r"$\dot{M}$", logy=LOG_MDOT)
+    bplt.diag_plot(ax_flux[1], diag, 'Phi_b', dump['t'], ylabel=r"$\phi_{BH}$", logy=LOG_PHI)
   elif movie_type == "radial":
 
     # TODO just record these in analysis output...
@@ -115,7 +115,7 @@ def plot(n):
     Ptot_r = eht_profile(geom, Pg + Pb, jmin, jmax)
     betainv_r = eht_profile(geom, Pb/Pg, jmin, jmax)
 
[…]
