### 2021-07-09 AFD-Illinois/iharm3d#30 @bprather
Hi,

Yeah, the grid sizing errors stems from a few bad experiences putting too few zones inside the event horizon (it should be causally isolated from the boundary, which due to the reconstruction stencil requires a few zones' buffer).

You can generate eht_out.p with 'python script/analysis/eht_analysis.py dumps/' -- that will generate a file named according to the simulation start and end times.  When parallelizing analysis, we merge a bunch of these files using eht_unify.py, but a single file can be safely renamed eht_out.p and used as input for movie.py.

Also added this to the README in #32.
