### 2021-07-13 a82ca90
updated readme for basic plotting
