### 2021-07-13 8a10959
added basic plotting script and associated param file
