+++ script/analysis/simple/basic_analysis.py
@@ -1,0 +1,468 @@
+######################################################################
+#
+# Simple analysis and plotting script to process problem output
+# Plots mhdmodes, bondi and torus problem (2D and 3D)
+#
+######################################################################
+
+
+import numpy as np
+import matplotlib
+matplotlib.use('Agg')
+import matplotlib.pyplot as plt
+from matplotlib import gridspec
+import os,psutil,sys
+import h5py
+from mpl_toolkits.axes_grid1 import make_axes_locatable
+import multiprocessing as mp
+
+# Parallelize analysis by spawning several processes using multiprocessing's Pool object
+def run_parallel(function,dlist,nthreads):
+    pool = mp.Pool(nthreads)
+    pool.map_async(function,dlist).get(720000)
+    pool.close()
+    pool.join()
+
+# Initialize global variables
+globalvars_keys = ['PROB','NDIMS','DUMPSDIR','PLOTSDIR']
+globalvars = {}
+grid ={}
+
+
+# Function to generate poloidal (x,z) slice
+# Argument must be variable, patch pole (to have x coordinate plotted correctly), averaging in phi option
+def xz_slice(var, patch_pole=False, average=False):
+	xz_var = np.zeros((2*grid['n1'],grid['n2']))
+	if average:
+		var = np.mean(var,axis=2)
+		for i in range(grid['n1']):
+			xz_var[i,:] = var[grid['n1']-1-i,:]
+			xz_var[i+grid['n1'],:] = var[i,:]
+	else:
+		angle = 0.; ind = 0
+		for i in range(grid['n1']):
+			xz_var[i,:] = var[grid['n1']-1-i,:,ind+grid['n3']//2]
+			xz_var[i+grid['n1'],:] = var[i,:,ind]
+	if patch_pole:
+		xz_var[:,0] = xz_var[:,-1] = 0
+	return xz_var
+
+
+# Function to generate poloidal (y,z) slice
+# Argument must be variable, patch pole (to have y coordinate plotted correctly), averaging in phi option
+# Not really called but can include a function call 
+def yz_slice(var, patch_pole=False, average=False):
+	yz_var = np.zeros((2*grid['n1'],grid['n2']))
+	if average:
+		var = np.mean(var,axis=2)
+		for i in range(grid['n1']):
+			yz_var[i,:] = var[grid['n1']-1-i,:]
+			yz_var[i+grid['n1'],:] = var[i,:]
+	else:
+		angle = np.pi/2; ind = np.argmin(abs(grid['phi'][0,0,:]-angle))
+		for i in range(grid['n1']):
+			yz_var[i,:] = var[grid['n1']-1-i,:,ind+grid['n3']//2]
+			yz_var[i+grid['n1'],:] = var[i,:,ind]
+	if patch_pole:
+		yz_var[:,0] = yz_var[:,-1] = 0
+	return yz_var
+
+
+# Function to generate toroidal (x,y) slice
+# Argument must be variable, averaging in theta option
+def xy_slice(var, average=False, patch_phi=False):
+    if average:
+        xy_var = np.mean(var,axis=1)
+    else:
+        xy_var = var[:,grid['n2']//2,:]
[…]
