+++ README.md
@@ -96,7 +96,7 @@ $python3 basic_analysis.py -p path/to/params_analysis/file
 ```
 The script by default parallelizes the analysis by using python's `multiprocessing` module. You can get around this by setting `nthreads` to `1` in main. For the 3D `torus` problem, it plots the density and plasma beta-inverse (magnetic pressure/gas pressure) in the XZ (poloidal) and XY (toroidal) plane. It overlays the poloidal density plot with magnetic field lines. For the 2D `torus` problem, it generates similar poloidal plots. If you're using the script on the output of a `bondi` problem, it will generate the poloidal density plot. Note that the `bondi` problem in `iharm3d` is unmagnetized and it wouldn't make sense to plot plasma beta-inverse. Finally, the script plots the density in XZ and XY plane for the `mhdmodes` problem.
 
-We hope that this script sheds some light on the way data is stored in the dump files and grid file (a more detailed summary can be found [here](https://github.com/AFD-Illinois/docs/wiki/GRMHD-Output-Format) and [here](https://github.com/AFD-Illinois/docs/wiki/Grid-Output-Format), and acts as a primer for the calculations performed to compute various qunatities of interest, and generate simple plots. 
+We hope that this script sheds some light on the way data is stored in the dump files and grid file (a more detailed summary can be found [here](https://github.com/AFD-Illinois/docs/wiki/GRMHD-Output-Format) and [here](https://github.com/AFD-Illinois/docs/wiki/Grid-Output-Format)), and acts as a primer for the calculations performed to compute various qunatities of interest, and generate simple plots. 
 
 If you're looking for a more complete set of scripts that calculates and plots a near-exhaustive list of relevant GRMHD diagnostics, have a look at [pyHARM](https://github.com/AFD-Illinois/pyHARM).
 
