### 2021-07-13 678068a
Make default torus parameters more normal
Previous defaults were for test scripts, which now set all relevant
parameters for themselves when running comparisons. No reason
to be needlessly obtuse any more.
