+++ prob/torus/parameters.h
@@ -9,13 +9,13 @@
 /* GLOBAL RESOLUTION */
 #define N1TOT 128
 #define N2TOT 128
-#define N3TOT 128
+#define N3TOT 1
 
 /* MPI DECOMPOSITION */
 /* COUNTERINTUITIVE: Split N3, N2, N1 order to keep k smaller than i,j*/
 #define N1CPU 1
-#define N2CPU 2
-#define N3CPU 2
+#define N2CPU 1
+#define N3CPU 1
 
 /* METRIC
  *   MINKOWSKI, MKS
