### 2021-07-13 5463a3e
Make the basic_analysis script a bit more general
