### 2021-07-13 0110f6f
typos in plotting_bfield_lines function
