### 2021-07-19 b1f98f5
Add a note on contributing to final section
