### 2021-07-26 66ef2a0
do away with most nested if statements
