+++ core/electrons.c
@@ -12,16 +12,15 @@
 
 // TODO put these in options with a default in decs.h
 // Defined as in decs.h, CONSTANT not included in ALLMODELS version
+// KAWAZURA is run by default if ALLMODELS=0 
 #define KAWAZURA  9
 #define WERNER    10
 #define ROWAN     11
 #define SHARMA    12
-#define CONSTANT 5
-#define FE_MODEL KAWAZURA
+#define CONSTANT 5 //tbh, this is never considered 
 
 void fixup_electrons_1zone(struct FluidState *S, int i, int j, int k);
 void heat_electrons_1zone(struct GridGeom *G, struct FluidState *Sh, struct FluidState *S, int i, int j, int k);
-double get_fel(struct GridGeom *G, struct FluidState *S, int i, int j, int k);
 double get_fels(struct GridGeom *G, struct FluidState *S, int i, int j, int k, int model);
 
 void init_electrons(struct GridGeom *G, struct FluidState *S)
@@ -33,14 +32,10 @@ void init_electrons(struct GridGeom *G, struct FluidState *S)
     // Initialize entropies
     S->P[KTOT][k][j][i] = (gam-1.)*S->P[UU][k][j][i]*pow(S->P[RHO][k][j][i],-gam);
 
-    // Initialize different models' entropies
-    #if ALLMODELS
+    // Initialize model entropy(ies)
     for (int idx = KEL0; idx < NVAR ; idx++) {
       S->P[idx][k][j][i] = (game-1.)*uel*pow(S->P[RHO][k][j][i],-game);
     }
-    #else
-    S->P[KEL][k][j][i] = (game-1.)*uel*pow(S->P[RHO][k][j][i],-game);
-    #endif
   }
 
   // Necessary?  Usually called right afterward
@@ -67,16 +62,11 @@ inline void heat_electrons_1zone(struct GridGeom *G, struct FluidState *Ss, stru
 
   //double uel = 1./(game-1.)*S->P[KEL][k][j][i]*pow(S->P[RHO][k][j][i],game);
 
-  // Evolve different models' entropies
-  #if ALLMODELS
+  // Evolve model entropy(ies)
   for (int idx = KEL0; idx < NVAR ; idx++) {
     double fel = get_fels(G, Ss, i, j, k, idx);
     Sf->P[idx][k][j][i] += (game-1.)/(gam-1.)*pow(Ss->P[RHO][k][j][i],gam-game)*fel*(kHarm - Sf->P[KTOT][k][j][i]);
   }
-  #else
-  double fel = get_fel(G, Ss, i, j, k);
-  Sf->P[KEL][k][j][i] += (game-1.)/(gam-1.)*pow(Ss->P[RHO][k][j][i],gam-game)*fel*(kHarm - Sf->P[KTOT][k][j][i]);
-  #endif
 
   // TODO bhlight calculates Qvisc here instead of this
   //double ugHat = S->P[KTOT][k][j][i]*pow(S->P[RHO][k][j][i],gam)/(gam-1.);
@@ -93,14 +83,13 @@ inline void heat_electrons_1zone(struct GridGeom *G, struct FluidState *Ss, stru
 }
 
 // New function for ALLMODELS runs.
-#if ALLMODELS
[…]
