+++ core/u_to_p.c
@@ -189,13 +189,9 @@ int U_to_P(struct GridGeom *G, struct FluidState *S, int i, int j, int k,
   S->P[U3][k][j][i] = (gamma/(W + Bsq))*(Qtcon[3] + QdB*Bcon[3]/W);
 
 #if ELECTRONS
-#if ALLMODELS
   for (int idx = KEL0; idx < NVAR ; idx++) {
     S->P[idx][k][j][i] = S->U[idx][k][j][i]/S->U[RHO][k][j][i];
   }
-#elif
-  S->P[KEL][k][j][i] = S->U[KEL][k][j][i]/S->U[RHO][k][j][i];
-#endif
   S->P[KTOT][k][j][i] = S->U[KTOT][k][j][i]/S->U[RHO][k][j][i];
 #endif // ELECTRONS
 
