### 2021-07-26 7cbf24e
referencing equations for e-heating models
