+++ core/electrons.c
@@ -89,6 +89,7 @@ inline double get_fels(struct GridGeom *G, struct FluidState *S, int i, int j, i
   double bsq = bsq_calc(S, i, j, k);
   double fel = 0.0;
 if (model == KAWAZURA) {
+	// Equation (2) in http://www.pnas.org/lookup/doi/10.1073/pnas.1812491116
   double Tpr = (gamp-1.)*S->P[UU][k][j][i]/S->P[RHO][k][j][i];
   double uel = 1./(game-1.)*S->P[model][k][j][i]*pow(S->P[RHO][k][j][i],game);
   double Tel = (game-1.)*uel/S->P[RHO][k][j][i];
@@ -103,9 +104,11 @@ if (model == KAWAZURA) {
   double QiQe = 35./(1. + pow(beta/15.,-1.4)*exp(-0.1/Trat));
   fel = 1./(1. + QiQe);
 } else if (model == WERNER) {
+	// Equation (3) in http://academic.oup.com/mnras/article/473/4/4840/4265350
   double sigma = bsq/S->P[RHO][k][j][i];
   fel = 0.25*(1+pow(((sigma/5.)/(2+(sigma/5.))), .5));
 } else if (model == ROWAN) {
+	// Equation (34) in https://iopscience.iop.org/article/10.3847/1538-4357/aa9380
   double pres = (gamp-1.)*S->P[UU][k][j][i]; // Proton pressure
   double pg = (gam-1)*S->P[UU][k][j][i];
   double beta = pres/bsq*2;
@@ -113,6 +116,7 @@ if (model == KAWAZURA) {
   double betamax = 0.25/sigma;
   fel = 0.5*exp(-pow(1-beta/betamax, 3.3)/(1+1.2*pow(sigma, 0.7)));
 } else if (model == SHARMA) {
+	// Equation for \delta on  pg. 719 (Section 4) in https://iopscience.iop.org/article/10.1086/520800
   double Tpr = (gamp-1.)*S->P[UU][k][j][i]/S->P[RHO][k][j][i];
   double uel = 1./(game-1.)*S->P[model][k][j][i]*pow(S->P[RHO][k][j][i],game);
   double Tel = (game-1.)*uel/S->P[RHO][k][j][i];
