### 2021-09-13 #3336 @arfon
> @arfon - just a ping on this, since it seems to have gone somewhat silent over the last two months

Thanks for the ping. I just emailed @bgiacoma to ask them to try and complete their review in the next couple of weeks.

:wave: @cpalenzuela – how are you getting along with your review here?
