### 2021-09-29 #3336 @bgiacoma
@arfon can you explain the Performance checklist? The authors provide a scaling plot in the paper, but what do you mean with "have they been confirmed"? Confirmed by the authors or confirmed by us?
