### 2021-09-29 #3336 @arfon
@bgiacoma – the point of this item is to check that if the authors are making claims about the performance of their software, have you (the reviewer) been able to verify them?
