### 2021-09-29 #3336 @bgiacoma
@arfon thanks for the reply. I have to say that I do not think I have the time to redo those scaling tests. They refer to 4 different problems with different resolutions. I would even need some help to know how to run some of them (e.g., parameters used).
