### 2021-09-29 f48e114
First pass at addressing some comments from Bruno Giacomazzo in the paper.
