+++ paper/paper.md
@@ -48,7 +48,7 @@ bibliography: paper.bib
 ---
 # `iharm3D` Functionality and Purpose
 
-`iharm3D`^[https://github.com/AFD-Illinois/iharm3d] is an open-source C code for simulating black hole accretion systems in arbitrary stationary spacetimes using ideal general-relativistic magnetohydrodynamics (GRMHD). It is an implementation of the HARM ("High Accuracy Relativistic Magnetohydrodynamics") algorithm outlined in @gammie_harm:_2003 with updates as outlined in @mckinney_measurement_2004 and @noble_primitive_2006.  The code is most directly derived from @ryan_bhlight:_2015 but but with radiative transfer portions removed.  HARM is a conservative finite-volume scheme for solving the equations of ideal GRMHD, a hyperbolic system of partial differential equations, on a logically Cartesian mesh in arbitrary coordinates.
+`iharm3D`^[https://github.com/AFD-Illinois/iharm3d] is an open-source C code for simulating black hole accretion systems in arbitrary stationary spacetimes using ideal general-relativistic magnetohydrodynamics (GRMHD). It is an implementation of the HARM ("High Accuracy Relativistic Magnetohydrodynamics") algorithm outlined in @gammie_harm:_2003 with updates as outlined in @mckinney_measurement_2004 and @noble_primitive_2006.  The code is most directly derived from @ryan_bhlight:_2015 but with radiative transfer portions removed.  HARM is a conservative finite-volume scheme for solving the equations of ideal GRMHD, a hyperbolic system of partial differential equations, on a logically Cartesian mesh in arbitrary coordinates.
 
 # Statement of Need
 
@@ -64,6 +64,8 @@ Multiple codes already exist for solving the ideal GRMHD equations on regular Eu
 - HARM-Noble (@noble_primitive_2006, @noble_direct_2009, @noble_circumbinary_2012, @zilhao_dynamic_2014, @bowen_quasi-periodic_2018)
 - IllinoisGRMHD (@etienne_illinoisgrmhd_2015)
 - KORAL (@sadowski_semi-implicit_2013, @sadowski_numerical_2014)
+- GRHydro (@mosta_grhydro_2014)
+- Spritz (@cipolletta_spritz_2020, @cipolletta_spritz_2021)
 
 The emphasis of `iharm3D` development is to provide a simple, fast, and scalable open update to the original open-source implementation of HARM (@gammie_harm:_2003).
 
@@ -73,14 +75,14 @@ In MHD, uncorrected discretization errors inevitably lead to violations of the n
 
 `iharm3D` also retains numerical evaluation of all metric-dependent quantities, allowing trivial modification of the coordinate system or background spacetime so long as the line element is available in analytic form.  This can be used as a form of static mesh refinement, since the coordinates can be adapted to place resolution in areas of interest (e.g., an accretion disk midplane).
 
-In GRMHD, "conserved" variables (energy and momentum density) are complicated analytic functions of "primitive" variables (density, pressure, and velocity).  Conserved variables are stepped forward in time and then inversion to primitives is done numerically. `iharm3d` uses the "$1D_W$" scheme outlined in @noble_primitive_2006.
[…]
