### 2021-09-29 f732ac3
Update wording around units & scale, add current status of electron heating models.
