+++ paper/paper.md
@@ -75,9 +75,9 @@ In MHD, uncorrected discretization errors inevitably lead to violations of the n
 
 `iharm3D` also retains numerical evaluation of all metric-dependent quantities, allowing trivial modification of the coordinate system or background spacetime so long as the line element is available in analytic form.  This can be used as a form of static mesh refinement, since the coordinates can be adapted to place resolution in areas of interest (e.g., an accretion disk midplane).
 
-In GRMHD, "conserved" variables (energy and momentum densities) are complicated analytic functions of "primitive" variables (density, pressure, and velocity).  Conserved variables are stepped forward in time and then inversion to primitives is done numerically. `iharm3d` uses the "$1D_W$" scheme outlined in @noble_primitive_2006.  As the equations of GRMHD are rescalable, any scale may be chosen for these variables.  For numerical stability, we choose units in which $GM = c = 1$, and scale the density such that the densest zone in the initial conditions has $\rho = 1$.  The mass and length are rescaled in post-processing to reflect the particular system under study.
+In GRMHD, "conserved" variables (energy and momentum densities) are complicated analytic functions of "primitive" variables (density, pressure, and velocity).  Conserved variables are stepped forward in time and then inversion to primitives is done numerically. `iharm3d` uses the "$1D_W$" scheme outlined in @noble_primitive_2006.  As the equations of ideal GRMHD are rescalable, any consistent set of units may be chosen to evolve them in the code.  For numerical stability, we choose units in which $GM = c = 1$, and scale the density such that the densest zone in the initial conditions has $\rho = 1$.  The mass and length are rescaled in post-processing to reflect the particular system under study.
 
-To model a collisionless plasma, `iharm3D` implements an optional scheme that provides a means of tracking and partitioning dissipation into ions and electrons (@ressler_electron_2015). The code implements the turbulent cascade models of @howes_prescription_2010 and @kawazura_thermal_2019, but new models are easy to implement and welcome.
+To model a collisionless plasma, `iharm3D` implements an optional scheme that provides a means of tracking and partitioning dissipation into ions and electrons (@ressler_electron_2015). Currently the code implements five heating models: those of @howes_prescription_2010, @kawazura_thermal_2019, @werner_non-thermal_2018, @rowan_electron_2017, and @sharma_electron_2007.
 
 To avoid catastrophic failures caused by discretization error, especially in low density regions, fluid variables are bounded at the end of each step. Typical `iharm3D` bounds in black hole accretion problems are enforced as follows:
 
