+++ paper/iharm3d_paper.bib
@@ -1484,7 +1484,7 @@ publicly available.},
 	month = jul,
 	year = {2020},
 	note = {ADS Bibcode: 2020CQGra..37m5010C},
-	keywords = {Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology, magnetohydrodynamics, neutron stars, numerical relativity},
+	keywords = {General Relativity and Quantum Cosmology, Astrophysics - High Energy Astrophysical Phenomena, magnetohydrodynamics, neutron stars, numerical relativity},
 	pages = {135010},
 	file = {Full Text PDF:/var/home/bprather/Zotero/storage/HZKVEETZ/Cipolletta et al. - 2020 - Spritz a new fully general-relativistic magnetohy.pdf:application/pdf},
 }
@@ -1503,7 +1503,7 @@ publicly available.},
 	month = apr,
 	year = {2021},
 	note = {ADS Bibcode: 2021CQGra..38h5021C},
-	keywords = {Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology, magnetohydrodynamics, neutron stars, numerical relativity},
+	keywords = {General Relativity and Quantum Cosmology, Astrophysics - High Energy Astrophysical Phenomena, magnetohydrodynamics, neutron stars, numerical relativity},
 	pages = {085021},
 	file = {Full Text PDF:/var/home/bprather/Zotero/storage/IE3MJUXS/Cipolletta et al. - 2021 - Spritz general relativistic magnetohydrodynamics .pdf:application/pdf},
 }
@@ -1525,4 +1525,58 @@ publicly available.},
 	keywords = {General Relativity and Quantum Cosmology},
 	pages = {015005},
 	file = {Full Text PDF:/var/home/bprather/Zotero/storage/JJGMYPX7/Mösta et al. - 2014 - GRHydro a new open-source general-relativistic ma.pdf:application/pdf},
+}
+
+@article{werner_non-thermal_2018,
+	title = {Non-thermal particle acceleration in collisionless relativistic electron-proton reconnection},
+	volume = {473},
+	issn = {0035-8711},
+	url = {https://ui.adsabs.harvard.edu/abs/2018MNRAS.473.4840W},
+	doi = {10.1093/mnras/stx2530},
[…]
