### 2021-09-30 #3336 @bgiacoma
> I apologize for the delay.
>
> A few initial comments about the paper:
>
> 1. in the abstract (lines 17 and 18) there is a typo: "but but"
> 2. In the statement of needs, please add references also to GRHydro (https://ui.adsabs.harvard.edu/abs/2014CQGra..31a5005M/abstract) and Spritz (https://ui.adsabs.harvard.edu/abs/2020CQGra..37m5010C/abstract and https://ui.adsabs.harvard.edu/abs/2021CQGra..38h5021C/abstract). They are both publicly available GRMHD codes.
> 3. in the implementation notes, please specify which units are used (I assume geometric?)
> 4. in the implementation notes, line 62, u is defined as the internal energy, but I think this is the internal energy density
> 5. figure 1, are you plotting the L1 norm of the quantities or the L1 norm of the error (i.e., the difference between the exact and the numerical solution)?

I think the paper is also missing a (short) description of the state of the field (i.e., Do the authors describe how this software compares to other commonly-used packages?).
