### 2021-09-30 #3336 @cpalenzuela
I also apologize for taking so long. Initially I could run quite easily the example given in the README. But when I tried to go beyond that example I found several issues. Among others, my main issues are:

1) I compiled the problem mhdmodes but when tried to run it failed with the following error message
./harm -p param.dat
Fatal error in PMPI_Cart_create: Invalid argument, error stack:
PMPI_Cart_create(340).....: MPI_Cart_create(MPI_COMM_WORLD, ndims=3, dims=0x7ffe9b9b6ee8, periods=0x7ffe9b9b6f40, reorder=1, comm_cart=0xa40e0c) failed
MPIR_Cart_create_impl(194): fail failed
MPIR_Cart_create(58)......: Size of the communicator (1) is smaller than the size of the Cartesian topology (16)
[unset]: write_line error; fd=-1 buf=:cmd=abort exitcode=1133068

2) I tried to run the torus problem in the Mare Nostrum cluster in order to double check the scaling of the code, but could not make it run in that infrastructure by just modifying my submit scripts. I got the following error in that case:
Fatal error in PMPI_Comm_rank: Invalid communicator, error stack:
PMPI_Comm_rank(122): MPI_Comm_rank(MPI_COMM_NULL, rank=0xa4684c) failed
PMPI_Comm_rank(75).: Null communicator
