### 2021-09-30 dc70211
Add a short summary of how iharm3D fits into the current GRMHD landscape
