+++ paper/paper.md
@@ -52,9 +52,9 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Numerical simulations are crucial in modeling observations of active galactic nuclei, such as the recent horizon-scale results from the Event Horizon Telescope (EHTC) and GRAVITY collaborations.  The computational simplicity of ideal GRMHD enables the generation of long, high-resolution simulations and broad parameter-exploration studies that can be compared to observations for parameter inference.
+Numerical simulations are crucial in modeling observations of active galactic nuclei, such as the recent horizon-scale results from the Event Horizon Telescope and GRAVITY collaborations.  The computational simplicity of ideal GRMHD enables the generation of long, high-resolution simulations and broad parameter-exploration studies that can be compared to observations for parameter inference.
 
-Multiple codes already exist for solving the ideal GRMHD equations on regular Eulerian meshes in 3D.  Some examples of codes currently in use:
+Multiple codes already exist for solving the ideal GRMHD equations on regular Eulerian meshes in 3D, including:
 
 - Athena++ (@stone_athena_2020, @white_extension_2016)
 - BHAC (@porth_black_2017)
@@ -69,21 +69,23 @@ Multiple codes already exist for solving the ideal GRMHD equations on regular Eu
 
 As the length of this list illustrates, the field of GRMHD simulation is now well established, and many codes now exist to serve different needs.  These codes can be distinguished by the trade-offs they make in prioritizing speed, simplicity, and generality, with the latter encompassing, e.g., support for dynamical spacetimes, adaptive mesh refinement, or higher-order integration schemes.
 
-In particular, `iharm3D` development focuses on providing the simplest and fastest possible code capable of simulating the original systems of interest when designing HARM, even at the cost of features aimed at more general applicability.  It provides a fast and scalable modern implementation of HARM, but maintains the conventions and simple structure of the original described in @gammie_harm:_2003.  The result is a code relatively easy to understand and modify, yet capable of running simulations at state-of-the-art scale.
+In particular, `iharm3D` aims to be a simple and fast code capable of simulating the original systems of interest when designing HARM, even at the cost of features aimed at more general applicability.  It provides a fast and scalable update to HARM, but maintains the conventions and structure of the original described in @gammie_harm:_2003.  The result is a code that is relatively easy to understand and modify, yet capable of running simulations at state-of-the-art scale.
 
 # Implementation Notes
 
 In MHD, uncorrected discretization errors inevitably lead to violations of the no-monopoles condition $\nabla \cdot B = 0$.  As in the original HARM implementation, `iharm3D` uses the "Flux-CT" scheme for cell-centered constrained transport outlined in @toth_b0_2000.
 
-`iharm3D` also retains numerical evaluation of all metric-dependent quantities, allowing trivial modification of the coordinate system or background spacetime so long as the line element is available in analytic form.  This can be used as a form of static mesh refinement, since the coordinates can be adapted to place resolution in areas of interest (e.g., an accretion disk midplane).
+`iharm3D` also retains numerical evaluation of all metric-dependent quantities, allowing trivial modification of the coordinate system or background spacetime so long as the line element is available in analytic form.  This can be used as a form of static mesh refinement, since the coordinates can be adapted to place resolution in areas of interest (e.g., near the accretion disk midplane).
 
[…]
