### 2021-10-01 65bfefc
Updated plot formatting, better descriptions of performance characteristics and tests
