### 2021-10-01 #3336 @bprather
The co-authors and I have updated the text to address the initial comments from @bgiacoma. The substantial changes are an addition to the Statement of Need describing the field in slightly more detail as the word limit allows, and a brief paragraph addressing units in Implementation Details.  We've also taken the opportunity to make minor updates to phrasing and formatting.

In particular, the text now answers the questions raised:
3. We use scaled units GM = c = 1 depending on the central object, so not strictly geometric
4. u is indeed the internal energy density
5. The convergence plots show the L1 norm of the error

@cpalenzuela, it sounds like both issues are related to the somewhat arcane way that iharm3D sets the number of MPI ranks it uses.  After compiling once, you can modify the header file in "build_archive/parameters.h" and change the N1CPU, N2CPU, and N3CPU parameters to your desired MPI communicator layout, then recompile.  The default MPI layouts for test problems were not always single-rank, for historical reasons related to automated testing that no longer apply.  As both reviewers have run across this counter-intuitive default now, we've fixed it in iharm3D PR [39](https://github.com/AFD-Illinois/iharm3d/pull/39)
