### 2021-10-05 #3336 @cpalenzuela
@bprather I can now compile and run all the tests locally, but I can not still run in the Mare Nostrum cluster, so I can not check the weak/strong scaling performance. Either I am doing something very wrong, or the code is not portable easily to other clusters/infrastructures. Maybe few more details about how to compile and run in other clusters than TACC would be useful.
