### 2021-10-05 #3336 @bprather
@cpalenzuela Sorry it's proven so difficult to get working -- as it was written specifically for KNL originally, easy portability wasn't our top-tier concern for iharm3D, as this review has highlighted (and helped us to address somewhat).
Part of the reason we don't have a generic cluster tutorial, though, is because everywhere we've tried iharm3D has proven so different.  MPI stacks seem to come in as many flavors as there are clusters, so while we've gotten it running on everything from a Raspberry Pi to OLCF Summit, every compile has proven at least a little unique.  If you open an issue at the [repo](https://github.com/AFD-Illinois/iharm3d/issues) with some details of the problem you're seeing, I'd be happy to help try to fix this.
