### 2021-10-07 #3336 @cpalenzuela
@bprather  I have been able to perform the scaling test and they look reasonable. I do not have any further issues with the code or with the paper.
