### 2021-10-08 22faccc
Print a clearer message when iharm3D is run with a communicator different from the compiled size.
