### 2021-10-13 df4cb1a
Bump version reflecting release.  Code as reviewed and updated for publication in JOSS.
