### 2021-10-13 #3336 @bprather
I've cut a release and uploaded the source to Zenodo [here](https://zenodo.org/record/5567372). The archive DOI is 10.5281/zenodo.5567372
